windows support, unicode decode fix, improvements
Made a few improvements. Started by adding Windows support, which inadvertently ended up fixing a lot of problems for linux and macos as well.
Changes:
- Windows support
Because cmd prompt and powershell do env variables differently, this required figuring out which one the user is using. Then changing the console outputs to reflect which cli the user is using. (this includes macos and linux, it always outputs the correct commanad)

-Removed pycookiecheat in favor of browser_cookie3
Initially browser_cookie3 was implemented for windows users since pycookiecheat is macos/linux only. However upon further investigation, the utf decode issues on macos and linux are caused by pycookiecheat. Changing the code to use browser_cookie3 fixed these issues

-code rewrite
Since the code now needs to check if the user in on windows and whether or not theyre using cmd prompt, it was easier to break down the code into multiple functions. This allows for much more control and cleaner output. 

-Exclusion Filter
There is now one variable to control what parts you want to remove from the final printout. Updating this as time goes on will make troubleshooting easier for everyone. Hopefully this solves the majority of "me me cookie no work" posts 

Just in case I forget, browser_cookie3 needs to be added to the requirements and the script in the repo's root directory also need to be changed.